+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.2
 
 - fix: reload
+- feat: currentRoute
 
 # 1.0.1